Tech: évite d’importer pikaday dans les tests unitaires
