+++ src/scripts/main.js
@@ -13,7 +13,8 @@ import SUIVI_IMAGES from '../suivi_*.svg'
 import { register as registerTimeAgo } from 'timeago.js'
 
 import { bindSuppressionTotale } from './actions'
-import { estPageThematique, pageThematique } from './page/thematiques/main'
+import { pageThematique } from './page/thematiques/main'
+import { estPageThematique } from './page/thematiques/navigation'
 import App from './app'
 
 var app = new App(SUIVI_IMAGES)
